updated delete category
