+++ TabloidFullStack/TabloidFullStack/client/tabloid/src/components/Categories/Category.js
