+++ TabloidFullStack/TabloidFullStack/client/tabloid/src/components/Categories/CategoryForm.js
@@ -1,31 +1,23 @@
-import { useEffect, useState } from "react";
-import { useNavigate, useParams } from "react-router-dom";
-import { editCategory, getCategoryById } from "../Managers/CategoryManager";
+import React, {useState, useContext } from "react";
 import { Form, FormGroup, Card, CardBody, Label, Input, Button } from "reactstrap";
+import { useNavigate } from "react-router-dom";
+import { addCategory } from "../../Managers/CategoryManager";
 
-export const EditCategory = () => {
+export const CategoryForm = () => {
     const [category, update] = useState({
         name: "",
-    });
+    })
 
     const navigate = useNavigate();
 
-    const { categoryId } = useParams();
-
-    useEffect(() => {
-        getCategoryById(categoryId)
-        .then((categoryArray) => {
-            update(categoryArray)
-        })
-    }, []);
-
     const handleSaveButtonClick = (event) => {
         event.preventDefault()
-        editCategory(category)
-        .then(() => {
-            navigate("/category")
-        })
-    }
+        const categoryToAPI = {
+            Name: category.name,
+        };
+        return addCategory(categoryToAPI)
+        .then(navigate("/category"));
+    };
 
     return (
         <div className="container pt-4">
@@ -34,12 +26,12 @@ export const EditCategory = () => {
                     <CardBody>
                         <Form>
                             <FormGroup>
-                                <Label for="name">Edit Category Name</Label>
+                                <Label for="name">Category Name</Label>
                                 <Input
                                     required autoFocus
                                     type="text"
                                     className="form-control"
-                                    placeholder="Enter a new name for your category"
+                                    placeholder="Enter the name of your category here"
                                     value={category.name}
                                     onChange={
                                         (event) => {
@@ -52,11 +44,13 @@ export const EditCategory = () => {
                             </FormGroup>
                         </Form>
                         <Button color="info" onClick={(clickEvent) => handleSaveButtonClick(clickEvent)}>
-                            SAVE
+                            SUBMIT
                         </Button>
                     </CardBody>
                 </Card>
             </div>
         </div>
-    )
-}
+    );
+};
+
+export default CategoryForm;
