+++ TabloidFullStack/TabloidFullStack/client/tabloid/src/components/Categories/EditCategory.js
@@ -1,23 +1,31 @@
-import React, {useState, useContext } from "react";
+import { useEffect, useState } from "react";
+import { useNavigate, useParams } from "react-router-dom";
+import { editCategory, getCategoryById } from "../../Managers/CategoryManager";
 import { Form, FormGroup, Card, CardBody, Label, Input, Button } from "reactstrap";
-import { useNavigate } from "react-router-dom";
-import { addCategory } from "../Managers/CategoryManager";
 
-export const CategoryForm = () => {
+export const EditCategory = () => {
     const [category, update] = useState({
         name: "",
-    })
+    });
 
     const navigate = useNavigate();
 
+    const { categoryId } = useParams();
+
+    useEffect(() => {
+        getCategoryById(categoryId)
+        .then((categoryArray) => {
+            update(categoryArray)
+        })
+    }, []);
+
     const handleSaveButtonClick = (event) => {
         event.preventDefault()
-        const categoryToAPI = {
-            Name: category.name,
-        };
-        return addCategory(categoryToAPI)
-        .then(navigate("/category"));
-    };
+        editCategory(category)
+        .then(() => {
+            navigate("/category")
+        })
+    }
 
     return (
         <div className="container pt-4">
@@ -26,12 +34,12 @@ export const CategoryForm = () => {
                     <CardBody>
                         <Form>
                             <FormGroup>
-                                <Label for="name">Category Name</Label>
+                                <Label for="name">Edit Category Name</Label>
                                 <Input
                                     required autoFocus
                                     type="text"
                                     className="form-control"
-                                    placeholder="Enter the name of your category here"
+                                    placeholder="Enter a new name for your category"
                                     value={category.name}
                                     onChange={
                                         (event) => {
@@ -44,13 +52,11 @@ export const CategoryForm = () => {
                             </FormGroup>
                         </Form>
                         <Button color="info" onClick={(clickEvent) => handleSaveButtonClick(clickEvent)}>
-                            SUBMIT
+                            SAVE
                         </Button>
                     </CardBody>
                 </Card>
             </div>
         </div>
-    );
-};
-
-export default CategoryForm;
+    )
+}
